requirements.txt
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
chart-studio==1.1.0
click==8.1.3
dash==2.7.0
dash-auth==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
Flask-Compress==1.13
Flask-SeaSurf==1.1.1
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
pandas==1.5.2
plotly==5.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-ldap==3.4.3
pytz==2022.6
requests==2.28.1
retrying==1.3.4
six==1.16.0
tenacity==8.1.0
ua-parser==0.16.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
